requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.3
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.4
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-tables2==2.6.0
django-widget-tweaks==1.5.0
numpy==1.25.2
packaging==23.1
pandas==2.1.0
plotly==5.16.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.7.1
tzdata==2023.3
